Don't use backticks in CoffeeScript
Backticks allow snippets of JavaScript to be embedded in CoffeeScript. While
some folks consider backticks useful in a few niche circumstances, they should
be avoided because so none of JavaScript's "bad parts", like <tt>with</tt> and
<tt>eval</tt>, sneak into CoffeeScript.

This can be enforced by Hound with the `no_backticks` option in CoffeeLint,
which is enabled by default.

http://www.coffeelint.org/#options

Also: alphabetize the CoffeeScript guidelines.